This was the problem of randomly failing logins:
Session with postgresql session storage didn't always save in time when a redirect was issued by the server. This has something to do with how Chrome handles redirects (doesn't read the whole response, but ditches the stream). This thread helped me to get on track:

https://github.com/voxpelli/node-connect-pg-simple/issues/31#issuecomment-230596077

Basically, when storing something to session and then doing redirects, we have to explicitly save the session and do the redirect in the save success callback.

:(