Fix copy values in table with odp
getFraValues used to filter out odps and after perttus changes this is no longer the case. instead the values are filtered out by years. hasFraValues was renamed to hasTableValues and still filters out odp's as it's only needed for knowing if the view has data in it's tables or not.