Fix the following scenario:
User invites herself to the platform with another google login email address.
User clicks the login link from her email while still logged in to the FRA platform with another google login email.
This caused the email used for the inviting process to get the new role for country and
invite was considered to be used.

Now the current user is always logged out when invitation UUID is part of the URL.