PR - #1746 - geo statistics (#1996)
* Show transparency only when the layer is checked (#1742)

Adding a conditional to show the layer transparency only if the layer
is checked, to save space.

* Add util to get the forest estimations from the API (#1746)

Added a new util to get the forest estimations in a tabular
manner, including the name of the source, the area and its
percentage relative to the total land area.

* Adding a table component with a download button (#1746)

The table component is intended to be used in the statistics panel.

* Adding a table to visualize the forest estimations (#1746)

Added a panel to visualize the forest estamations data in a table,
including the option to download it.

* Adding a graph panel to visualize forest estimations (#1746)

Added a panel with a chart that shows the area per source and
its percentage relative to the land area. It also includes a
bar with the forest area and its percentage relative to the total
land area.

Co-authored-by: Yhoan Alejandro Guzman Garcia <yhoan.alejandro@Yhoans-MacBook-Pro.local>