1747 add functionality to load google satellite as base map (#2111)
* Add satellite view as base map and map view control (#1747)

Adding the hybrid satellite view (with labels) as the default,
and adding google's buttons to switch between satellite and
roadmap views.

* Fix bug crashing the geomap on safari (#1747)

The regexp to format the areas was causing the site to crash
on Safari. It is now replaced with the Numbers util to avoid that
and to have consistent format in the application.

* Add formatting to messages in the statistics panel

Adding a better format to the messages displayed when
the statistics are loading or return an error.

* Update forest estimations URL call (#1747)

* Add conditional linking to the country selector (#1747)

If the user changes the country while in the geo page, they
should remain in the geo page. If the user selects a region, then
they will be redirected to the dashboard.

* Re-center the map and adjust zoom when changing countries

Adding the functionality to focus on a country when using the
country selector and the data from the 'bounds' endpoint.
The default center and zoom correspond to Filand.

---------

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>