Display precalculated recipe data when available (2038)
Adding the recipe layer data to the forest estimations
statistics when available. This was done by refactoring
the fetch Forest Estimations function, turning it into
its own hook, which calls the API endpoint only when the
country or the year (hardcoded as of now) change. When
there is a precalculated recipe available, it is added
to the table data without calling the API endpoind again.