PR - #2038 - add pre-calculated recipe estimation to statistics (#2144)
* Add hook to keep track of precalculated recipies (2038)

Adding a hook to keep track to when a precalculated recipe
is available for the current forest layer set up.

* Add custom estimation metada (2038)

Adding custom estimation metadata and types for the
extra estimations displayed (besides forest area products).

* Display precalculated recipe data when available (2038)

Adding the recipe layer data to the forest estimations
statistics when available. This was done by refactoring
the fetch Forest Estimations function, turning it into
its own hook, which calls the API endpoint only when the
country or the year (hardcoded as of now) change. When
there is a precalculated recipe available, it is added
to the table data without calling the API endpoind again.

* Add missing forest estimations interface properties (2038)

---------

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>