Add and migrate logic to handle layers on countryIso change
Migrated to its own hook the logic that handles the changes on
the layers, and added the logic to handle the reload when the
country ISO switches. That was done by adding two more states to
the ForestOptions: pedingLayers and failedLayers. By managing
those states it is possible to reload the layers for the new
country without unnecesary calls to the API. That was not
possible before since the code was not implemented to
handle multiple layers being fetched at once.

The two new states also open the posibility to add a loading
or failed status in the UI, giving more feedback to the user.