Add the layers handler at GeoMap level
Added the layers handler to the GeoMap component given that the
user can switch countries without having the Map Visualizer Panel
Opended, which is where the logic was previously.

Also added a default center to the configuration. This is because
the map will not load unless it has a default center or the function
'panTo' is called to a location. It was the case that sometimes the
API call that fetches the center of the country would take a couple of
seconds, and since the map did not have a center, it resulted in the
UI just showing a grey box.