table 4a sub category validation fix
Table 4a, New Zealand why are all the cells for the “…of which” categories highlighted? The validation check should not consider cells that are empty. If all sub-categories are reported on then the sum should be equal to the parent, but in case of missing values this is not the case.