2137 reload layers when country changes (#2205)
* Add and migrate logic to handle layers on countryIso change

Migrated to its own hook the logic that handles the changes on
the layers, and added the logic to handle the reload when the
country ISO switches. That was done by adding two more states to
the ForestOptions: pedingLayers and failedLayers. By managing
those states it is possible to reload the layers for the new
country without unnecesary calls to the API. That was not
possible before since the code was not implemented to
handle multiple layers being fetched at once.

The two new states also open the posibility to add a loading
or failed status in the UI, giving more feedback to the user.

* Add the layers handler at GeoMap level

Added the layers handler to the GeoMap component given that the
user can switch countries without having the Map Visualizer Panel
Opended, which is where the logic was previously.

Also added a default center to the configuration. This is because
the map will not load unless it has a default center or the function
'panTo' is called to a location. It was the case that sometimes the
API call that fetches the center of the country would take a couple of
seconds, and since the map did not have a center, it resulted in the
UI just showing a grey box.

* Format graph label to be consistent with the table format

* Fix: set opacity when country ISO changes

Previously, the opacities selected at the moment where
the country ISO changed were not set again for the new
layers. That is now fixed by keeping track of the value
of the opacities, avoiding adding that dependency since
it would trigger useEffects too often.

---------

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>