Feature/1748 geo backgrounds (#2217)
* Add sepal layer

* Modify Sepal body to avoid memory error

* Update Mosaic controls based on a call with Eugenia

* Move mosaic controls to a separate component

This should improve performance because the mosaic component is now only
loaded when 'Background' is clicked

* Fix unnecessary Sepal call

* Separate UI and applied mosaic options

UI options are reflected in the UI. Applied options are the ones actually
used when fetching the background.

* Make comment more describing

* Enable apply changes button if options have changed

* Apply mosaic changes

* Set Sentinel request body based on mosaic options

* Use Landsat as default because it's faster

* Refactor styling

* Style mosaic checkboxes

* Fix gray map

* Change component file names to uppercase

* Fix old method name

* Set mosaic country dynamically

* Switch mosaic on the fly if country is switched

* Move mosaic state up

The issue before was that it required the Background popup to be open.
Now the popup can be open or closed and the mosaics still load.

* Remove log comment

---------

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>