GEO: refactor params, load user's asset (#2435)
* add custom forest source

* refactor request params move to body

* refactor entry name

* refactor params to LayerSource

* refactor schemes

* refactor endpoints to post

* Fetch forest layers using POST & Agreement layer status

Refactoring to fetch the layers with post after the backend change,
and adding the logic to handle the fetch status of the agreement layer.

* Fix custom asset validation

* Adjust layer inputs styling

* Add support for fetching a custom layer

Adding an input where the user can specify a custom
asset id to be displayed in the map. Adding the redux
states and the handling in the corresponding hooks.

* add LayerSatus

---------

Co-authored-by: yaguzmang <alejandroguzman.dev@gmail.com>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>