2415 allow user to calculate agreement area (#2482)
* Migrate MapContext to use global mapController instead

Getting rid of the MapContext and its provider so the map
instance can be accessed from outside of the GeoMap component.
This change is introduced because the layers control will be
migrated to the navigation component (outside the map).
Redux can not be used in this case since the map object from the
google library is not serializable, and adding it would be an
anti-pattern. Instead, there is now a global mapController that holds
the reference to the google map when it is set, and there also is a
redux state that stores whether the map is available or not.

* Move mosaic property to mosaic options in redux

* Move the loading spinner to its own file

* Delete duplicated enum 'LayerStatus'

* Refactor Geo statistics to use redux

Refactoring the geo statistics to store the data using redux,
which will make it easier to later modify the data from other
places in the UI.

* Fetch forest estiation based on agreement layers

Fetching a new estimation and refactoring the statistics graph to
update when the data changes.