2455 add protected areas section (#2490)
* Divide the logic to handle map layers

Dividing the logic to handle different kinds of layers,
since there will be more in the future. Having them in
separate files is better for readability and makes the
code more modular (e.g. you could now just handle forest
layers on their own, without having to handle all layers).

* Rename "layers" to "forestLayers" & add protected area layers

* Rename redux reducer & add ProtectedAreasOptions

Renaming redux reducers to be more specific, considering
that there will be more layers that are not part of the
forest layers. Also, adding the types for the protected
areas layers.

* Add redux state management for Protected Areas

* Add UI logic for the Protected Areas

* Prevent custom asset id input from reseting to empty

* Clear unused scss classes and fix ui bug in input

Clearing the unsed styling, and using global variables instead.
Also fixed a bug causing the method trim() to be called on a null
value, which is the initial value of the customAssetId.

---------

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>