Comment out layer section components
Previously, the way the layers worked in the ui
was that each type (Forest, Burned Area, Protected
Area) would have its own component, but that is no
longer the case as we are migrating to a generic
approach. The upcoming UI component will render all
types of layers, thus the hardcoded components need
to be removed.