Render layers when toggling and when fetching sucessfully
A layer is fetched using postLayer when it is selected and it does not have
a mapId set. Then when the status of the fetch is successful, the layer is
rendered on the map. If it is loading or has failed, the remove method is
called to ensure it is not shown. If the fetch fails, the layer is
unselected and the mapId is set to empty so the user can try again.