Add TreeCoverPercentageControl component
Added a component to handle layers with a percent picker.
Added logic to handle that component in LayerSectionPanel. The
new logic includes a 'fetch' flag to 'toggle' layer, so we can prevent
fetching the layer when toggling. This is because when the user selects
a layer, they should be able to see its menu before it is fetched.