Geo refactor/handle gte tree cover percent layers (#2626)
* Rename minTreeCoverPercentage to gteTreeCoverPercent

Renaming to gteTreeCoverPercent as that's how it is called
in the backend.

* Add redux logic to set gteTreeCoverPercent

* Add TreeCoverPercentageControl component

Added a component to handle layers with a percent picker.
Added logic to handle that component in LayerSectionPanel. The
new logic includes a 'fetch' flag to 'toggle' layer, so we can prevent
fetching the layer when toggling. This is because when the user selects
a layer, they should be able to see its menu before it is fetched.

* Add custom hook to fetch layer when gteTreeCoverPercent changes

This custom hook fetches a layer that has gteCoverPercent whenever it
changes. This hook intentionally does not fetch the layer when the
countryIso changes. This will be done later with a custom hook of its own.