Geo refactor/handle agreement layer (#2633)
* Add LayerResponseData type and logic

Adding a response data type since the layers API not only
returns a mapId but also some other information, like palette,
which is necessary for the agreement layer.

* Add logic to build the agreement layer req body

Adding a builder for the agreement layer calls, using the
section state with which the postLayer action was called. That
state is later used to get the selected layers and add them to
the layers array in the api call.

* Add redux state management for the agreement layer

Adding logic to set the agreement level, and to set the palette
once the data is retrieved using the postLayer action.

* Add custom hook to fetch agreement layer on level change

* Add UI components for the agreement layer

Adding the UI components and logic for the agreement layer. The
styles are the same as before. I replaced some scss colors with
the defined generics we have, but not all colors were replaceable.
Further styling changes can be later added in a styling specific refactor.

* Get LayerState and SectionState from within postLayer action

* Update useAgreementLevel

Deleting the unnecessary hooks and updating after delegating the state
check to the postLayer action.

* Move and rename useAgreementLevel to useFetchAgreementLevelLayer

Moving useAgreementLevel (now useFetchAgreementLevelLayer, as that is
more accurate and descriptive) to GeoMap/hooks.

* Move and rename useGteTreeCoverPercent to useFetchGteTreeCoverPercentLayer

Moving this hook to GeoMap/hooks and giving it a better descriptive name as well.

* Delete unnecessary usePrevious from useFetchGteTreeCoverPercentLayer

Since we now get the state from within postLayer, the usePrevious hook here
becomes unnecessary.