Extend useFetchGteTreeCoverPercentLayer to handle other options
The useFetchGteTreeCoverPercentLayer hook was set up to keep track
of the selected percent. That logic is exactly the same as the logic
needed to handle the selected year, so now the hook is extended to
accept different options properties. The agreement layer needs more
specific logic so it is omitted.