Update layer recipes type and comment out deprecated function
The recipe type used the deprecated 'gteHansenTreeCoverPerc' property
and an array of layer keys. That was not sufficient because the
layers now have other options such as assetId or year. Instead, recipes
now consist of an array of 'LayerSource', which encapsulates all the
options.

The function 'getRecipeAgreementAreaProperty' was used for the statistics
module, and after the changes to the recipes it no longer works. It will be
fixed if needed after the statistics module is addressed.