Geo refactor/recipe selector (#3315)
* Update layer recipes type

The recipe type used the deprecated 'gteHansenTreeCoverPerc' property
and an array of layer keys. That was not sufficient because the
layers now have other options such as assetId or year. Instead, recipes
now consist of an array of 'LayerSource', which encapsulates all the
options.

The function 'getRecipeAgreementAreaProperty' was used for the statistics
module, and after the changes to the recipes, it no longer works. It will be
fixed if needed after the statistics module is addressed.

* Add optional layerSource param to postLayer

Adding the option to fetch a layer with a layerSource,
this way if the layerSource is given, then there won't be a
need to construct it.

* Add state management for recipe name and all layerOptions

Adding a recipe property to GeoState to keep track of the
selected recipe. Also added a reducer to update all options
in a layer.

* Add 'setLayerSectionRecipe' action

This action selects all layers that are part of
a recipe, with their given options (e.g. gteTreeCoverPercent,
year, etc.). It also unselects all layers that are not part of
the recipe.

* Add hook to keep track of a section's selected recipe

* Add recipe selector UI component

* Turn toggleLayer slice action into an asyncThunk action

* Delegate cache logic to toggleLayer action

* Turn layer opacity slice actions into async thunks

Adding setLayerOpacity and setSectionGlobalOpacity async thunks
to handle cases where a layer opacity is set but the layer has not
been fetched (e.g after selecting a recipe).